Accidentally working in wrong folder, fixed to correct folder
This is going to be one massive code dump because I accidentally setup git in the wrong folder. This adds a class, weapon, armos, leveling, and spellcasting system. It also enables some editor tools so that users can quickly create some sample content that can either be used directly or as references for creating more assets.